Convert Plugin\Runner from abstract class to interface
